requirements.txt
selenium>=4.15.0
webdriver-manager>=4.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
pandas>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
fake-useragent>=1.4.0
undetected-chromedriver>=3.5.0
geopy>=2.4.0
ratelimit>=2.2.0
backoff>=2.2.0
loguru>=0.7.0
pydantic>=2.5.0
asyncpg>=0.28.0
flask>=2.3.0
flask-cors>=4.0.0
setuptools>=80.0.0